Create README.md
Unit 2 | Assignment - The VBA of Wall Street

Background

You are well on your way to becoming a programmer and Excel master! In this homework assignment you will use VBA scripting to analyze real stock market data. Depending on your comfort level with VBA, choose your assignment from Easy, Moderate, or Hard below.

Easy:

Create a script that will loop through each year of stock data and grab the total amount of volume each stock had over the year.
You will also need to display the ticker symbol to coincide with the total volume.
Your result should look as follows (note: all solution images are for 2015 data).

Moderate:

Create a script that will loop through all the stocks and take the following info.

Yearly change from what the stock opened the year at to what the closing price was.
The percent change from the what it opened the year at to what it closed.
The total Volume of the stock
Ticker symbol

You should also have conditional formatting that will highlight positive change in green and negative change in red.

CHALLENGE:

Make the appropriate adjustments to your script that will allow it to run on every worksheet just by running it once.
This can be applied to any of the difficulties.